+++ __test__/CustomizeProfileModal.test.tsx
@@ -5,7 +5,7 @@ import { render, waitFor, screen, fireEvent } from '@testing-library/react';
 import { notifications } from '@mantine/notifications';
 import { DataProvider } from '@/src/contexts/DataContext';
 import ProfilePage from '@/src/components/Profile/ProfilePage';
-import { storeImage } from '@/src/components/utils/s3Helpers/UserProfilePictureS3Helper';
+import { storeImage } from '@/src/utils/s3Helpers/UserProfilePictureS3Helper';
 
 function createMockImageFile(filename = 'test-image.jpg', type = 'image/jpeg') {
   return new File([new Blob([], { type })], filename);
