+++ src/hooks/communityCustomHooks.tsx
@@ -7,7 +7,7 @@ import {
   getAllCommunityDetailsAPI,
 } from '../api/services/community';
 import { getCurrentUserID } from './usersCustomHooks';
-import { retrieveImage as retrieveProfilePicture } from '@/src/components/utils/s3Helpers/UserProfilePictureS3Helper';
+import { retrieveImage as retrieveProfilePicture } from '@/src/utils/s3Helpers/UserProfilePictureS3Helper';
 
 export const communityUpdateSubject = new Subject<void>();
