+++ src/hooks/eventsCustomHooks.tsx
@@ -10,8 +10,8 @@ import {
   deleteEventAPI,
 } from '../api/services/event';
 import { Event, Visibility } from '@/src/types/types';
-import { retrieveImage as retrieveProfilePicture } from '@/src/components/utils/s3Helpers/UserProfilePictureS3Helper';
-import { retrieveImage as retrieveEventImage } from '@/src/components/utils/s3Helpers/EventImageS3Helper';
+import { retrieveImage as retrieveProfilePicture } from '@/src/utils/s3Helpers/UserProfilePictureS3Helper';
+import { retrieveImage as retrieveEventImage } from '@/src/utils/s3Helpers/EventImageS3Helper';
 
 export const useCreateEvent = () => {
   const handleCreateEvent = async (eventData: any) => {
