+++ src/hooks/marketplaceCustomHooks.tsx
@@ -10,8 +10,8 @@ import {
   listUserSavedListingsAPI,
 } from '../api/services/marketplace';
 import { ItemForSale, Visibility } from '@/src/types/types';
-import { retrieveImage as retrieveProfilePicture } from '@/src/components/utils/s3Helpers/UserProfilePictureS3Helper';
-import { retrieveImage as retrieveItemImage } from '@/src/components/utils/s3Helpers/ItemForSaleImageS3Helper';
+import { retrieveImage as retrieveProfilePicture } from '@/src/utils/s3Helpers/UserProfilePictureS3Helper';
+import { retrieveImage as retrieveItemImage } from '@/src/utils/s3Helpers/ItemForSaleImageS3Helper';
 
 export const useCreateListing = () => {
   const handleCreateListing = async (itemData: any) => {
