+++ src/hooks/postsCustomHooks.tsx
@@ -12,7 +12,7 @@ import {
 import { getCurrentUser, getCurrentUserID } from './usersCustomHooks';
 import { Post, CommentDataInput, PostDataInput, Visibility, CommentItem } from '@/src/types/types';
 import { getCurrentCommunityID } from './communityCustomHooks';
-import { retrieveImage as retrieveProfilePicture } from '@/src/components/utils/s3Helpers/UserProfilePictureS3Helper';
+import { retrieveImage as retrieveProfilePicture } from '@/src/utils/s3Helpers/UserProfilePictureS3Helper';
 
 export const useCreatePost = () => {
   const [error, setError] = useState<string | undefined>();
