+++ src/utils/communityUtils.ts
