+++ src/utils/relevantCommunitiesHelpers/getClosestCommunities.ts
@@ -4,13 +4,13 @@ import { generateClient } from '@aws-amplify/api';
 import { listCommunities } from '@/src/graphql/queries';
 
 // You might need to replace `configureAmplify` depending on how your Amplify is set up.
-import { configureAmplify } from '@/app/api/utils/amplifyServerConfig';
+import { configureAmplify } from '@/src/components/ConfigureAmplifyServer';
 import { Community } from '@/src/API';
 
 // Type for the community with distance
 export interface CommunityWithDistance {
-    community: Community;
-    distanceKm: number;
+  community: Community;
+  distanceKm: number;
 }
 
 configureAmplify();
@@ -28,8 +28,10 @@ export async function getClosestCommunities(coordinates: string): Promise<Commun
 
   try {
     const distancePromises = items.map(async (item: Community) => {
-        const response = await fetch(`/api/getDistance?origin=${encodeURIComponent(coordinates)}&destination=${encodeURIComponent(item.coordinates)}`);
-        const data = await response.json();
+      const response = await fetch(
+        `/api/getDistance?origin=${encodeURIComponent(coordinates)}&destination=${encodeURIComponent(item.coordinates)}`
+      );
+      const data = await response.json();
       if (data?.error) throw new Error(`Failed to fetch communities: ${data.error}`);
 
       return {
