+++ src/utils/s3Helpers/UserProfilePictureS3Helper.tsx
@@ -2,7 +2,7 @@ import { uploadData } from '@aws-amplify/storage';
 
 import ConfigureAmplifyClientSide from '@/src/components/ConfigureAmplify';
 import { getUserAPI, updateUserProfilePicAPI } from '@/src/api/services/user';
-import { deleteImageFromS3, retrieveImageURLFromS3 } from '../utilFunctions';
+import { deleteImageFromS3, retrieveImageURLFromS3 } from '../s3Utils';
 
 ConfigureAmplifyClientSide();
